% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_abundance_plot.R
\name{el_abundance_plot}
\alias{el_abundance_plot}
\title{General 16S Abundance Plot}
\usage{
el_abundance_plot(dir, threshold = 100, taxtextsize = 3,
  labeltextsize = 12)
}
\arguments{
\item{dir}{Directory containing the uparse output files (otu table in biom format, fasta of otu reps, taxonomy table, phylogenetic tree). Should be string ie "~/Downloads"}

\item{threshold}{Minimum number of sequences required for labeling taxons at species level}

\item{taxtextsize}{Font size of labeled taxons}

\item{labeltextsize}{Font size of sample labels on the x axis}
}
\description{
This function reads uparse output files from the specified directory and generates a 16S abundance plot at the species level
}
