% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erictools.R
\docType{package}
\name{erictools}
\alias{erictools}
\alias{erictools-package}
\title{Eric's tools}
\description{
Eric's personal function library for data analysis
}
\details{
To install/update this package, run: \code{devtools::install_github("elittmann/erictools")}
}
\author{
Eric Littmann

Maintainer: Eric Littmann
}
