% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.stack.el.R
\name{gg.stack.el}
\alias{gg.stack.el}
\title{old ggstack function}
\usage{
gg.stack.el(..., heights = NULL, gg.extras = NULL, gap = 0, margin = 1,
  units = "inches", as.list = FALSE)
}
\description{
ggstack function that still works on older R version
}
