% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{gm_mean}
\alias{gm_mean}
\title{CLR transformation function}
\usage{
gm_mean(x, na.rm = TRUE)
}
\description{
Performs centered log ratio transformation to normalize count data
}
\examples{
t <- get.otu.melt(phy)
t \%>\%
group_by(sample) \%>\%
summarize(pctseqs=clr(pctseqs))
}
