% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_readin.R
\name{metabolomics_readin}
\alias{metabolomics_readin}
\title{metabolomics read in function
This function reads in directory of csv files, using the filenames for method. Can handle multiple methods from multiple patients at the same time
Look in given directory for .csv files, read all into a dataframe, adding method name and cleaning up sample ID. Replace NA values with min/10 per compound or with given value
Requires dplyr, tidyr, stringr, reshape2}
\usage{
metabolomics_readin(directory, na.value = 1, min_over10 = F)
}
\arguments{
\item{directory}{directory containing .csv metabolomic tables in wide format}

\item{na.value}{if no value present for a compound, replace with this value}

\item{min_over10}{when TRUE will override na.value and replace na's with minimum value of the compound / 10}
}
\description{
metabolomics read in function
This function reads in directory of csv files, using the filenames for method. Can handle multiple methods from multiple patients at the same time
Look in given directory for .csv files, read all into a dataframe, adding method name and cleaning up sample ID. Replace NA values with min/10 per compound or with given value
Requires dplyr, tidyr, stringr, reshape2
}
