% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_hm.R
\name{plot_meta_hm}
\alias{plot_meta_hm}
\title{plot metabolomics heatmap
designed for use with gg.stack with other patient metadata, 16S, etc using 'Sample_ID' on the x axis}
\usage{
plot_meta_hm(data, row_mean_centered = T, organize_by = "method")
}
\arguments{
\item{data}{dataframe containing all metabolite data in long form for one patient. should have the following variables: Sample_ID, compound, peak, method}

\item{row_mean_centered}{whether to center by mean peak area per compound, default = TRUE}

\item{organize_by}{either facet by method or order by hierarchical cluster. options are: "method" or "cluster"}
}
\description{
plot metabolomics heatmap
designed for use with gg.stack with other patient metadata, 16S, etc using 'Sample_ID' on the x axis
}
\examples{
meta_plot <- plot_meta_hm(meta,row_mean_centered=T,organize_by="cluster")
gg.stack(tax,meta.plot,heights=c(1,5))
}
