% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_yt_tax.R
\name{plot_yt_tax}
\alias{plot_yt_tax}
\title{General 16S Abundance Plot
This function reads uparse output files from the specified directory and generates a 16S abundance plot at the species level}
\usage{
plot_yt_tax(phylo, pctseqs = T, threshold = 0.1, taxtextsize = 3,
  labeltextsize = 12)
}
\arguments{
\item{phylo}{phyloseq object (can use read_uparse to get uparse outputs into phyloseq...)}

\item{pctseqs}{When TRUE will return plots with % abundance, when false will return plots using number of sequences not normalized to 1}

\item{threshold}{Minimum number of sequences (numseqs) or % abundance (pctseqs) required for labeling taxons at species level}

\item{taxtextsize}{Font size of labeled taxons}

\item{labeltextsize}{Font size of sample labels on the x axis}
}
\description{
General 16S Abundance Plot
This function reads uparse output files from the specified directory and generates a 16S abundance plot at the species level
}
