% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uparse.R
\name{read_uparse}
\alias{read_uparse}
\title{uparse readin function
read outputs from Ying's uparse pipeline into a phyloseq object}
\usage{
read_uparse(directory)
}
\arguments{
\item{directory}{directory containing uparse output files (usually called 'uparse')
requires phyloseq, ape, yingtools2, dplyr}
}
\description{
uparse readin function
read outputs from Ying's uparse pipeline into a phyloseq object
}
\examples{
phy <- read_uparse("Z:/Eric/uparse")
}
