% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legend.R
\name{extract_legend}
\alias{extract_legend}
\title{Extract Legend Function
This function extracts a legend from a ggplot object, returning a ggplot object}
\usage{
extract_legend(a.gplot)
}
\description{
Extract Legend Function
This function extracts a legend from a ggplot object, returning a ggplot object
}
